update HomePage update FakerUtils
