+++ src/main/resources/localization/homepage_vi.properties
@@ -10,16 +10,3 @@ text.currency=VND
 
 input.from=T??
 input.to=?ê?n
-
-month.Jan=Tháng Mô?t
-month.Feb=Tháng Hai
-month.Mar=Tháng Ba
-month.Apr=Tháng T?
-month.May=Tháng N?m
-month.Jun=Tháng Sáu
-month.Jul=Tháng B?y
-month.Aug=Tháng Tám
-month.Sep=Tháng Chín
-month.Oct=Tháng M??i
-month.Nov=Tháng M??i M?t
-month.Dec=Tháng M??i Hai
